restructuración de carpetas y mejora de algoritmosalgorim
