+++ algoritmos/algoritmos_iniciales/matrices.py
@@ -1,0 +1,46 @@
+[[0.0, 15.73166666666667, 19.65166666666666, 2.543333333333333, 7.55, 3.76, 15.83833333333333, 22.38833333333333, 8.563333333333333, 7.185, 4.903333333333333, 10.02833333333333, 12.33666666666667, 5.098333333333333, 17.525, 2.943333333333333, 15.54833333333333, 15.105, 23.855, 23.7, 6.905], 
+ [15.73166666666667, 0.0, 7.505, 15.79833333333333, 17.24166666666667, 17.31666666666667, 1.248333333333334, 20.7, 21.325, 13.50333333333333, 18.17, 14.42333333333333, 6.694999999999999, 15.89166666666667, 20.06166666666667, 14.02333333333333, 26.15333333333333, 14.80333333333333, 14.235, 21.90833333333333, 20.38166666666667], 
+ [19.65166666666666, 7.505, 0.0, 19.285, 0.0, 0.0, 0.0, 22.07666666666666, 19.98166666666667, 17.37833333333333, 19.54833333333334, 15.8, 11.54166666666667, 17.27, 20.405, 18.5, 27.53166666666667, 16.18166666666666, 13.02333333333333, 18.08, 20.87333333333333], 
+ [2.543333333333333, 15.79833333333333, 19.285, 0.0, 8.533333333333333, 1.878333333333333, 16.52166666666666, 22.86333333333333, 7.165, 7.013333333333334, 5.986666666666666, 10.50166666666667, 10.76833333333333, 6.181666666666667, 18.0, 2.465, 16.54666666666667, 15.58, 24.32833333333333, 23.03, 7.4],
+ [7.55, 17.24166666666667, 0.0, 8.533333333333333, 0.0, 9.711666666666668, 0.0, 0.0, 0.0, 11.05833333333333, 5.12, 8.793333333333333, 18.555, 7.391666666666667, 16.29166666666667, 10.73833333333333, 9.998333333333333, 13.87166666666667, 22.62, 28.09333333333333, 7.938333333333333], 
+ [3.76, 17.31666666666667, 0.0, 1.878333333333333, 9.711666666666668, 0.0, 17.06833333333333, 22.66833333333333, 6.033333333333333, 8.281666666666666, 6.233333333333333, 10.30833333333333, 11.315, 7.038333333333333, 17.80666666666667, 3.733333333333333, 16.24, 15.38666666666667, 23.87666666666667, 21.89833333333333, 6.268333333333334],
+ [15.83833333333333, 1.248333333333334, 0.0, 16.52166666666666, 0.0, 17.06833333333333, 0.0, 20.82666666666666, 20.98833333333333, 13.63166666666667, 18.29833333333334, 14.55, 0.0, 0.0, 0.0, 14.15, 26.28166666666667, 14.93166666666667, 14.36333333333333, 21.19333333333333, 20.51],
+ [22.38833333333333, 20.7, 22.07666666666666, 22.86333333333333, 0.0, 22.66833333333333, 20.82666666666666, 0.0, 25.13333333333333, 16.505, 20.51, 15.50333333333333, 20.85, 18.23166666666667, 10.55833333333333, 18.38, 22.96333333333333, 15.53333333333333, 20.25, 30.10166666666667, 23.61833333333333], 
+ [8.563333333333333, 21.325, 19.98166666666667, 7.165, 0.0, 6.033333333333333, 20.98833333333333, 25.13333333333333, 0.0, 13.97666666666667, 6.815, 12.04833333333333, 15.17333333333333, 9.086666666666668, 19.54666666666667, 9.428333333333335, 13.025, 17.12666666666667, 25.875, 25.26666666666667, 1.841666666666667], 
+ [7.185, 13.50333333333333, 17.37833333333333, 7.013333333333334, 11.05833333333333, 8.281666666666666, 13.63166666666667, 16.505, 13.97666666666667, 0.0, 9.693333333333333, 5.411666666666666, 10.275, 0.0, 13.88, 0.0, 18.62333333333333, 10.88666666666667, 20.20833333333333, 26.28333333333333, 11.38833333333333],
+ [4.903333333333333, 18.17, 19.54833333333334, 5.986666666666666, 5.12, 6.233333333333333, 18.29833333333334, 20.51, 6.815, 9.693333333333333, 0.0, 7.295, 14.58666666666667, 3.641666666666667, 14.79333333333333, 6.989999999999999, 13.4, 12.37333333333333, 21.12166666666667, 23.97333333333334, 5.476666666666667],
+ [10.02833333333333, 14.42333333333333, 15.8, 10.50166666666667, 8.793333333333333, 10.30833333333333, 14.55, 15.50333333333333, 12.04833333333333, 5.411666666666666, 7.295, 0.0, 12.97333333333333, 7.765, 10.435, 8.22, 18.02666666666666, 9.49, 18.24, 28.98166666666667, 13.64166666666667],
+ [12.33666666666667, 6.694999999999999, 11.54166666666667, 10.76833333333333, 18.555, 11.315, 0.0, 20.85, 15.17333333333333, 10.275, 14.58666666666667, 12.97333333333333, 0.0, 13.83833333333333, 17.495, 8.76, 23.76333333333333, 13.95333333333333, 19.335, 19.16333333333333, 15.12], 
+ [5.098333333333333, 15.89166666666667, 17.27, 6.181666666666667, 7.391666666666667, 7.038333333333333, 0.0, 18.23166666666667, 9.086666666666668, 0.0, 3.641666666666667, 7.765, 13.83833333333333, 0.0, 0.0, 0.0, 0.0, 13.59166666666667, 22.34166666666667, 25.035, 8.538333333333332], 
+ [17.525, 20.06166666666667, 20.405, 18.0, 16.29166666666667, 17.80666666666667, 0.0, 10.55833333333333, 19.54666666666667, 13.88, 14.79333333333333, 10.435, 17.495, 0.0, 0.0, 16.26166666666667, 20.20666666666667, 10.115, 22.35833333333333, 32.21166666666667, 19.73], 
+ [2.943333333333333, 14.02333333333333, 18.5, 2.465, 10.73833333333333, 3.733333333333333, 14.15, 18.38, 9.428333333333335, 0.0, 6.989999999999999, 8.22, 8.76, 0.0, 16.26166666666667, 0.0, 17.05166666666667, 14.80166666666667, 23.55, 24.42, 8.408333333333333], 
+ [15.54833333333333, 26.15333333333333, 27.53166666666667, 16.54666666666667, 9.998333333333333, 16.24, 26.28166666666667, 22.96333333333333, 13.025, 18.62333333333333, 13.4, 18.02666666666666, 23.76333333333333, 0.0, 20.20666666666667, 17.05166666666667, 0.0, 21.36333333333333, 30.11166666666667, 29.705, 10.94833333333333], 
+ [15.105, 14.80333333333333, 16.18166666666666, 15.58, 13.87166666666667, 15.38666666666667, 14.93166666666667, 15.53333333333333, 17.12666666666667, 10.88666666666667, 12.37333333333333, 9.49, 13.95333333333333, 13.59166666666667, 10.115, 14.80166666666667, 21.36333333333333, 0.0, 19.87166666666667, 31.86666666666667, 18.75666666666667], 
+ [23.855, 14.235, 13.02333333333333, 24.32833333333333, 22.62, 23.87666666666667, 14.36333333333333, 20.25, 25.875, 20.20833333333333, 21.12166666666667, 18.24, 19.335, 22.34166666666667, 22.35833333333333, 23.55, 30.11166666666667, 19.87166666666667, 0.0, 25.87833333333333, 23.79833333333334], 
+ [23.7, 21.90833333333333, 18.08, 23.03, 28.09333333333333, 21.89833333333333, 21.19333333333333, 30.10166666666667, 25.26666666666667, 26.28333333333333, 23.97333333333334, 28.98166666666667, 19.16333333333333, 25.035, 32.21166666666667, 24.42, 29.705, 31.86666666666667, 25.87833333333333, 0.0, 17.02], 
+ [6.905, 20.38166666666667, 20.87333333333333, 7.4, 7.938333333333333, 6.268333333333334, 20.51, 23.61833333333333, 1.841666666666667, 11.38833333333333, 5.476666666666667, 13.64166666666667, 15.12, 8.538333333333332, 19.73, 8.408333333333333, 10.94833333333333, 18.75666666666667, 23.79833333333334, 17.02, 0.0]]
+
+#-------------------------------------------------KILOMETROS--------------------------------------------------------------------
+[
+[0.0, 7.5625, 15.5365, 1.1998, 4.7145, 1.7407, 7.9408, 17.1947, 4.2933, 3.2659, 2.1866, 6.0225, 5.447, 2.2133, 11.1505, 1.5775, 10.8288, 9.1456, 20.4871, 22.1445, 3.6114],
+[7.5625, 0.0, 3.3838, 7.743300000000001, 14.572, 8.523700000000002, 0.4847, 13.7974, 10.1522, 7.152100000000001, 14.8113, 10.1049, 2.6961, 13.4907, 18.0835, 7.0275, 19.8218, 8.273700000000002, 9.636899999999999, 19.1038, 10.7361],
+[15.5365, 3.3838, 0.0, 12.5438, 0.0, 0.0, 0.0, 16.0355, 13.912, 13.0649, 17.0494, 12.343, 5.0114, 15.7289, 17.9217, 9.6824, 22.0599, 10.5118, 7.7574, 16.5997, 13.9021],
+[1.1998, 7.743300000000001, 12.5438, 0.0, 5.072100000000001, 0.9118999999999999, 7.579800000000001, 17.4095, 3.5781, 3.3451, 2.8532, 6.233, 4.7117, 2.8799, 11.361, 1.3127, 11.1926, 9.3561, 20.7019, 21.1518, 3.3673],
+[4.7145, 14.572, 0.0, 5.072100000000001, 0.0, 4.8187, 0.0, 0.0, 0.0, 7.217, 2.6253, 6.8738, 9.106399999999999, 3.6476, 12.0019, 5.434699999999999, 5.287199999999999, 9.9969, 21.3384, 25.5947, 4.5417],
+[1.7407, 8.523700000000002, 0.0, 0.9118999999999999, 4.8187, 0.0, 7.8866, 20.0472, 2.906, 4.0899, 2.889, 8.875, 5.0185, 3.2185, 14.003, 2.0575, 11.4008, 11.9981, 19.7467, 20.4797, 2.6952],
+[7.9408, 0.4847, 0.0, 7.579800000000001, 0.0, 7.8866, 0.0, 13.9893, 10.1499, 7.344, 15.0032, 10.2968, 0.0, 0.0, 0.0, 7.219399999999999, 20.0138, 8.4656, 9.828899999999999, 17.1722, 10.928],
+[17.1947, 13.7974, 16.0355, 17.4095, 0.0, 20.0472, 13.9893, 0.0, 19.6963, 12.9161, 16.1829, 12.3127, 13.5906, 14.8623, 5.9742, 14.6038, 22.6587, 8.9984, 15.5515, 34.1368, 20.8703],
+[4.2933, 10.1522, 13.912, 3.5781, 0.0, 2.906, 10.1499, 19.6963, 0.0, 6.8679, 3.5313, 10.1024, 7.722899999999999, 4.5536, 15.2304, 4.8355, 9.7565, 13.2255, 24.567, 22.9781, 1.0494],
+[3.2659, 7.152100000000001, 13.0649, 3.3451, 7.217, 4.0899, 7.344, 12.9161, 6.8679, 0.0, 4.7283, 2.3061, 4.9896, 0.0, 8.2196, 0.0, 10.8948, 5.305, 17.5604, 21.1407, 5.9516],
+[2.1866, 14.8113, 17.0494, 2.8532, 2.6253, 2.889, 15.0032, 16.1829, 3.5313, 4.7283, 0.0, 6.5845, 8.2105, 1.6099, 11.7126, 3.397, 7.7031, 9.7077, 21.0492, 22.484, 2.912],
+[6.0225, 10.1049, 12.343, 6.233, 6.8738, 8.875, 10.2968, 12.3127, 10.1024, 2.3061, 6.5845, 0.0, 6.4451, 5.9051, 5.6418, 4.2199, 12.2362, 5.1248, 16.4706, 22.5962, 10.2374],
+[5.447, 2.6961, 5.0114, 4.7117, 9.106399999999999, 5.0185, 0.0, 13.5906, 7.722899999999999, 4.9896, 8.2105, 6.4451, 0.0, 10.0708, 10.0675, 4.2809, 15.2069, 6.286899999999999, 12.4538, 17.7187, 7.9895],
+[2.2133, 13.4907, 15.7289, 2.8799, 3.6476, 3.2185, 0.0, 14.8623, 4.5536, 0.0, 1.6099, 5.9051, 10.0708, 0.0, 0.0, 0.0, 0.0, 8.603299999999999, 19.9448, 23.8969, 4.0306],
+[11.1505, 18.0835, 17.9217, 11.361, 12.0019, 14.003, 0.0, 5.9742, 15.2304, 8.2196, 11.7126, 5.6418, 10.0675, 0.0, 0.0, 10.6147, 21.7754, 5.1865, 18.4207, 37.006, 15.6084],
+[1.5775, 7.0275, 9.6824, 1.3127, 5.434699999999999, 2.0575, 7.219399999999999, 14.6038, 4.8355, 0.0, 3.397, 4.2199, 4.2809, 0.0, 10.6147, 0.0, 11.3728, 8.927700000000002, 20.2735, 21.0351, 4.155399999999999],
+[10.8288, 19.8218, 22.0599, 11.1926, 5.287199999999999, 11.4008, 20.0138, 22.6587, 9.7565, 10.8948, 7.7031, 12.2362, 15.2069, 0.0, 21.7754, 11.3728, 0.0, 14.8006, 26.1421, 30.6412, 8.5221],
+[9.1456, 8.273700000000002, 10.5118, 9.3561, 9.9969, 11.9981, 8.4656, 8.9984, 13.2255, 5.305, 9.7077, 5.1248, 6.286899999999999, 8.603299999999999, 5.1865, 8.927700000000002, 14.8006, 0.0, 14.7105, 24.3734, 13.2923],
+[20.4871, 9.636899999999999, 7.7574, 20.7019, 21.3384, 19.7467, 9.828899999999999, 15.5515, 24.567, 17.5604, 21.0492, 16.4706, 12.4538, 19.9448, 18.4207, 20.2735, 26.1421, 14.7105, 0.0, 24.1215, 22.0368],
+[22.1445, 19.1038, 16.5997, 21.1518, 25.5947, 20.4797, 17.1722, 34.1368, 22.9781, 21.1407, 22.484, 22.5962, 17.7187, 23.8969, 37.006, 21.0351, 30.6412, 24.3734, 24.1215, 0.0, 14.8282],
+[3.6114, 10.7361, 13.9021, 3.3673, 4.5417, 2.6952, 10.928, 20.8703, 1.0494, 5.9516, 2.912, 10.2374, 7.9895, 4.0306, 15.6084, 4.155399999999999, 8.5221, 13.2923, 22.0368, 14.8282, 0.0]
+]
