+++ algoritmos/caso1/README.md
@@ -1,0 +1,1 @@
+- Algoritmo genético de optimización de rutas
