+++ algoritmos/caso2/README.md
@@ -1,0 +1,1 @@
+- Algoritmo genético de optimización de rutas reduciendo la flota de vehículos a la mitad
