--- datos/raw_data/prediccion/pedidos_2025.csv
@@ -1,21 +1,0 @@
-cliente,order_demand
-Cliente_1,725.0
-Cliente_2,721.0
-Cliente_3,768.0
-Cliente_4,743.0
-Cliente_5,679.0
-Cliente_6,705.0
-Cliente_7,713.0
-Cliente_8,741.0
-Cliente_9,707.0
-Cliente_10,690.0
-Cliente_11,753.0
-Cliente_12,692.0
-Cliente_13,726.0
-Cliente_14,712.0
-Cliente_15,740.0
-Cliente_16,726.0
-Cliente_17,708.0
-Cliente_18,720.0
-Cliente_19,694.0
-Cliente_20,848.0
